app/backend/requirements.txt
fastapi>=0.111.0
uvicorn[standard]>=0.29.0
sqlmodel>=0.0.18
pydantic>=2.7.0
pydantic-settings>=2.2.0
python-dotenv>=1.0.0
httpx>=0.27.0
openai>=1.30.0
python-dateutil>=2.9.0
Pillow>=10.3.0
pyzbar>=0.1.9
python-multipart>=0.0.9
